Fixed import of actions from extension profiles, now they don't overwrite all previous actions in the action provider.
git-svn-id: svn+ssh://svn.zope.org/repos/main/CMF/branches/1.6@41382 62d5b8a3-27da-0310-9561-8e5933582275

+++ CMFCore/exportimport/actions.py
@@ -122,11 +122,12 @@ class ActionsToolXMLAdapter(XMLAdapterBase):
             if provider_id not in self.context.listActionProviders():
                 self.context.addActionProvider(provider_id)
 
-            # delete any actions that are auto-created
-            provider = getToolByName(self.context, provider_id)
-            num_actions = len(provider.listActions())
-            if num_actions:
-                provider.deleteActions(range(0,num_actions))
+            if self.environ.shouldPurge():
+                # Delete provider's actions
+                provider = getToolByName(self.context, provider_id)
+                num_actions = len(provider.listActions())
+                if num_actions:
+                    provider.deleteActions(range(0, num_actions))
 
             # BBB: for CMF 1.5 profiles
             self._initOldstyleActions(child)
@@ -163,6 +164,12 @@ class ActionsToolXMLAdapter(XMLAdapterBase):
                         break  # only one permission is allowed
                     if permission:
                         break
+
+            # Remove previous action with same id and category.
+            old = [i for (i, action) in enumerate(provider.listActions())
+                   if action.id == action_id and action.category == category]
+            if old:
+                provider.deleteActions(old)
 
             provider.addAction(action_id, title, url_expr,
                                condition_expr, permission,
